refactor(canon): split modules over the source-length limit
Six modules crossed the 350-line source-length ratchet on this branch. Move
cohesive units into sibling modules without changing behavior: the import
specifier scanner, declaration-emit option types, the .vue.ts shim rewriter,
VirtualProject configuration accessors, tsconfig path remapping, and the
workspace-package incremental test.
